Fix: incorrect dataloader return value

+++ helper/dataloader.py
@@ -41,7 +41,7 @@ class TrajectoryDataset(Dataset):
         self.samples: List[Tuple[torch.Tensor, torch.Tensor]] = []
         fpaths = [data_dir / f"{id}.txt" for id in dataframe["unique_id"].values]
         for fpath, meta in zip(fpaths, metas):
-            data = torch.from_numpy(np.loadtxt(fpath))
+            data = torch.from_numpy(np.loadtxt(fpath)).float()
             segs = segment(data) if segment else [data]
             self.samples += [
                 (seg, meta)
